+++ apps/builder/src/app/space/[workspaceId]/instagrams/[id]/capi/page.tsx
@@ -1,85 +1,17 @@
-import {
-  metaConversionsService,
-  platformCredentialService,
-  workspaceService,
-} from "@chatbotx.io/business"
-import {
-  debugToken,
-  hasInstagramManageEventsScope,
-  toAppAccessToken,
-} from "@chatbotx.io/integration-instagram-facebook"
-import { notFound } from "next/navigation"
-import { checkMessagingAdsConnectionState } from "@/features/ads-campaign/queries"
-import { InstagramCapiTab } from "@/features/integration-instagram/components/instagram-capi-tab"
-import { findIntegrationInstagram } from "@/features/integration-instagram/queries"
+import { notFound, redirect } from "next/navigation"
 import { withWorkspaceIdAndIdSchema } from "@/features/workspaces/schema/resource"
-import { resolveOwnerForWorkspace } from "@/lib/platform-credential-owner"
 
-export default async function InstagramCapiPage(props: {
+/**
+ * The tab route moved from `/capi` to `/ads` — it is no longer CAPI-only,
+ * it also hosts the messaging-ads box and ads optimization. Kept as a
+ * redirect (not deleted) so old bookmarks/links to `.../capi` keep working.
+ */
+export default async function InstagramCapiRedirect(props: {
   params: Promise<{ workspaceId: string; id: string }>
 }) {
   const { data } = withWorkspaceIdAndIdSchema.safeParse(await props.params)
   if (!data) {
     return notFound()
   }
-
-  const { workspaceId, id } = data
-  const [workspace, integrationInstagram, messagingAdsConnectionState] =
-    await Promise.all([
-      workspaceService.findById({ id: workspaceId }),
-      findIntegrationInstagram({ workspaceId, id }),
-      checkMessagingAdsConnectionState({
-        workspaceId,
-        channel: "instagram",
-        integrationId: id,
-      }),
-    ])
-  const instagramCredential =
-    integrationInstagram.type === "facebook"
-      ? await platformCredentialService.resolveForOwner({
-          ownerId: await resolveOwnerForWorkspace(workspace),
-          type: "instagramFacebook",
-        })
-      : null
-  // A manual access token + dataset is ready without the OAuth scope, and a
-  // user-disconnected integration must stay disconnected — both skip the scope
-  // refresh (which can throw CapiScopeRefreshError on an expired OAuth token).
-  const usesManualToken = Boolean(
-    integrationInstagram.capiAccessToken && integrationInstagram.datasetId,
-  )
-  const capiDisconnected = Boolean(integrationInstagram.capiDisconnectedAt)
-  const refreshed =
-    instagramCredential && !(usesManualToken || capiDisconnected)
-      ? await metaConversionsService
-          .refreshCapiScopeCache({
-            channel: "instagram",
-            integration: integrationInstagram,
-            checkScope: async ({ accessToken }) => {
-              const debug = await debugToken({
-                inputToken: accessToken,
-                appAccessToken: toAppAccessToken(instagramCredential.config),
-                version: instagramCredential.config.version,
-              })
-              return hasInstagramManageEventsScope(debug.scopes)
-            },
-          })
-          .catch(() => integrationInstagram)
-      : integrationInstagram
-
-  const resolved = refreshed ?? integrationInstagram
-
-  return (
-    <InstagramCapiTab
-      capiDisconnected={capiDisconnected}
-      credentialAvailable={Boolean(instagramCredential)}
-      hasManualCapiAccessToken={Boolean(resolved.capiAccessToken)}
-      integrationInstagram={{
-        id: resolved.id,
-        type: resolved.type,
-        hasCapiScope: resolved.hasCapiScope,
-        datasetId: resolved.datasetId,
-      }}
-      messagingAdsConnectionState={messagingAdsConnectionState}
-    />
-  )
+  return redirect(`/space/${data.workspaceId}/instagrams/${data.id}/ads`)
 }
